sizeof on multiple initializer

+++ hello_world/6-size.c
@@ -6,9 +6,9 @@
 #include <stdio.h>
 int main(void)
 {
-printf("Size of a char: %lu byte(s)", sizeof(char));
-printf("Size of a int: %lu byte(s)", sizeof(int));
-printf("Size of a long: %lu byte(s)", sizeof(long));
-printf("Size of a long long: %lu byte(s)", sizeof(long long));
-printf("Size of a float: %lu byte(s)", sizeof(float));
+printf("Size of a char: %lu byte(s)", (unsigned long)sizeof(char));
+printf("Size of a int: %lu byte(s)", (unsigned long)sizeof(int));
+printf("Size of a long: %lu byte(s)", (unsigned long)sizeof(long));
+printf("Size of a long long: %lu byte(s)", (unsigned long)sizeof(long long));
+printf("Size of a float: %lu byte(s)", (unsigned long)sizeof(float));
 }
